Extract common
