+++ common/pom.xml
@@ -25,7 +25,7 @@
     <packaging>jar</packaging>
 
     <properties>
-        <flink.version>1.11.1</flink.version>
+        <flink.version>1.9.0</flink.version>
     </properties>
 
     <dependencies>
