+++ common/src/main/java/org/apache/flink/streaming/connectors/pulsar/internal/PulsarDeserializationSchema.java
@@ -18,9 +18,7 @@ import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
 import org.apache.flink.util.Collector;
-
 import org.apache.pulsar.client.api.Message;
-
 import java.io.IOException;
 import java.io.Serializable;
 
@@ -32,18 +30,6 @@ import java.io.Serializable;
  */
 @PublicEvolving
 public interface PulsarDeserializationSchema<T> extends Serializable, ResultTypeQueryable<T> {
-    /**
-     * Initialization method for the schema. It is called before the actual working methods
-     * {@link #deserialize} and thus suitable for one time setup work.
-     *
-     * <p>The provided {@link DeserializationSchema.InitializationContext} can be used to access additional features such as e.g.
-     * registering user metrics.
-     *
-     * @param context Contextual information that can be used during initialization.
-     */
-    default void open(DeserializationSchema.InitializationContext context) throws Exception{
-
-    }
 
     /**
      * Method to decide whether the element signals the end of the stream. If
