+++ common/src/main/java/org/apache/flink/streaming/connectors/pulsar/internal/ReaderThread.java
@@ -23,7 +23,6 @@ import org.apache.pulsar.client.impl.MessageIdImpl;
 import org.apache.pulsar.client.impl.conf.ClientConfigurationData;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -44,14 +43,13 @@ public class ReaderThread<T> extends Thread {
     protected final ExceptionProxy exceptionProxy;
     protected final String topic;
     protected final MessageId startMessageId;
+    private boolean failOnDataLoss = true;
 
     protected volatile boolean running = true;
 
-    private final PulsarDeserializationSchema<T> deserializer;
+    protected final PulsarDeserializationSchema<T> deserializer;
 
     protected volatile Reader<?> reader = null;
-
-    private boolean failOnDataLoss = true;
 
     public ReaderThread(
             PulsarFetcher owner,
@@ -117,22 +115,16 @@ public class ReaderThread<T> extends Thread {
     }
 
     protected void createActualReader() throws org.apache.pulsar.client.api.PulsarClientException, ExecutionException {
-        Map<String, Object> readerConf0 = new HashMap<>();
-        readerConf.entrySet().stream().forEach(entry -> {
-            if (!PulsarOptions.FAIL_ON_DATA_LOSS_OPTION_KEY.equals(entry.getKey())) {
-                readerConf0.put(entry.getKey(), entry.getValue());
-            }
-        });
         reader = CachedPulsarClient
                 .getOrCreate(clientConf)
                 .newReader()
                 .topic(topic)
                 .startMessageId(startMessageId)
                 .startMessageIdInclusive()
-                .loadConf(readerConf0)
+                .loadConf(readerConf)
                 .create();
         log.info("Create a reader at topic {} starting from message {} (inclusive) : config = {}",
-            topic, startMessageId, readerConf0);
+            topic, startMessageId, readerConf);
     }
 
     protected void skipFirstMessageIfNeeded() throws org.apache.pulsar.client.api.PulsarClientException {
@@ -142,8 +134,7 @@ public class ReaderThread<T> extends Thread {
         if (!startMessageId.equals(MessageId.earliest)
                 && !startMessageId.equals(MessageId.latest)
                 && ((MessageIdImpl) startMessageId).getEntryId() != -1) {
-            MessageIdImpl lastMessageId = (MessageIdImpl) this.owner.getMetadataReader().getLastMessageId(reader.getTopic());
-
+            MessageIdImpl lastMessageId = (MessageIdImpl) this.owner.getMetaDataReader().getLastMessageId(reader.getTopic());
             if (!messageIdRoughEquals(startMessageId, lastMessageId) && !reader.hasMessageAvailable()) {
                 MessageIdImpl startMsgIdImpl = (MessageIdImpl) startMessageId;
                 long startMsgLedgerId = startMsgIdImpl.getLedgerId();
@@ -151,17 +142,16 @@ public class ReaderThread<T> extends Thread {
 
                 // startMessageId is bigger than lastMessageId
                 if (startMsgLedgerId > lastMessageId.getLedgerId()
-                        || (startMsgLedgerId == lastMessageId.getLedgerId() && startMsgEntryId > lastMessageId.getEntryId())) {
-                    if (failOnDataLoss) {
-                        log.error("the start message id is beyond the last commit message id, with topic:{}", reader.getTopic());
-                        throw new RuntimeException("start message id beyond the last commit");
-                    } else {
-                        log.info("reset message to valid offset {}", startMessageId);
-                        this.owner.getMetadataReader().resetCursor(reader.getTopic(), startMessageId);
-                    }
+                    || (startMsgLedgerId == lastMessageId.getLedgerId() && lastMessageId.getEntryId() != -1
+                        && startMsgEntryId > lastMessageId.getEntryId())) {
+                    log.error("the start message id is beyond the last commit message id, with topic:{}, " +
+                            "start msgId:{}, last msgId:{}", reader.getTopic(), startMessageId, lastMessageId);
                 }
+                log.warn("reset message to valid offset");
+                this.owner.getMetaDataReader().resetCursor(reader.getTopic(), startMessageId);
+            } else {
+                failOnDataLoss = false;
             }
-
             while (currentMessage == null && running) {
                 currentMessage = reader.readNext(pollTimeoutMs, TimeUnit.MILLISECONDS);
                 if (failOnDataLoss) {
@@ -174,6 +164,7 @@ public class ReaderThread<T> extends Thread {
                         topic));
             } else {
                 currentId = currentMessage.getMessageId();
+                state.setOffset(currentId);
                 if (!messageIdRoughEquals(currentId, startMessageId) && failOnDataLoss) {
                     reportDataLoss(
                             String.format(
@@ -212,7 +203,11 @@ public class ReaderThread<T> extends Thread {
         this.running = false;
 
         if (reader != null) {
-            reader.close();
+            try {
+                reader.close();
+            } catch (IOException e) {
+                log.error("failed to close reader. ", e);
+            }
         }
 
         this.interrupt();
@@ -228,8 +223,10 @@ public class ReaderThread<T> extends Thread {
                 new IllegalStateException(message + PulsarOptions.INSTRUCTION_FOR_FAIL_ON_DATA_LOSS_TRUE));
     }
 
-    // used to check whether starting position and current message we got actually are equal
-    // we neglect the potential batchIdx deliberately while seeking to MessageIdImpl for batch entry
+    /** used to check whether starting position and current message we got actually are equal
+     * we neglect the potential batchIdx deliberately while seeking to MessageIdImpl for batch entry.
+     *
+     */
     public static boolean messageIdRoughEquals(MessageId l, MessageId r) {
         if (l == null || r == null) {
             return false;
