+++ pulsar-flink-1.11/src/main/java/org/apache/flink/streaming/connectors/pulsar/internal/PulsarDeserializationSchemaWrapper.java
@@ -37,7 +37,7 @@ public class PulsarDeserializationSchemaWrapper<T> implements PulsarDeserializat
         this.deserializationSchema = deserializationSchema;
     }
 
-    @Override
+//    @Override
     public void open(DeserializationSchema.InitializationContext context) throws Exception {
         this.deserializationSchema.open(context);
     }
