+++ pulsar-flink-1.11/src/test/java/org/apache/flink/streaming/connectors/pulsar/FlinkPulsarTableITest.java
@@ -16,9 +16,11 @@ package org.apache.flink.streaming.connectors.pulsar;
 
 import org.apache.flink.streaming.api.datastream.DataStreamSource;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
+import org.apache.flink.streaming.connectors.pulsar.internal.IncompatibleSchemaException;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarMetadataReader;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarOptions;
 import org.apache.flink.streaming.connectors.pulsar.internal.SchemaUtils;
+import org.apache.flink.streaming.connectors.pulsar.internal.SimpleSchemaTranslator;
 import org.apache.flink.streaming.connectors.pulsar.testutils.FailingIdentityMapper;
 import org.apache.flink.streaming.connectors.pulsar.testutils.SingletonStreamSink;
 import org.apache.flink.table.api.Table;
@@ -205,13 +207,14 @@ public class FlinkPulsarTableITest extends PulsarTestBaseWithFlink {
                 flList.subList(0, flList.size() - 1).stream().map(Objects::toString).collect(Collectors.toList()));
     }
 
-    private TableSchema getTableSchema(String topicName) throws PulsarClientException, PulsarAdminException, SchemaUtils.IncompatibleSchemaException {
+    private TableSchema getTableSchema(String topicName) throws PulsarClientException, PulsarAdminException,
+            IncompatibleSchemaException {
         Map<String, String> caseInsensitiveParams = new HashMap<>();
         caseInsensitiveParams.put(TOPIC_SINGLE_OPTION_KEY, topicName);
         PulsarMetadataReader reader = new PulsarMetadataReader(adminUrl, new ClientConfigurationData(), "", caseInsensitiveParams, -1, -1);
         SchemaInfo pulsarSchema = reader.getPulsarSchema(topicName);
-        FieldsDataType fieldsDataType = SchemaUtils.pulsarSourceSchema(pulsarSchema, false);
-        return SchemaUtils.toTableSchema(fieldsDataType);
+        final SimpleSchemaTranslator schemaTranslator = new SimpleSchemaTranslator();
+        return schemaTranslator.pulsarSchemaToTableSchema(pulsarSchema);
     }
 
     @Test
