+++ pulsar-flink-1.9/src/main/java/org/apache/flink/streaming/connectors/pulsar/FlinkPulsarRowSource.java
@@ -19,12 +19,11 @@ import org.apache.flink.streaming.api.functions.AssignerWithPeriodicWatermarks;
 import org.apache.flink.streaming.api.functions.AssignerWithPunctuatedWatermarks;
 import org.apache.flink.streaming.api.operators.StreamingRuntimeContext;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarDeserializationSchema;
+import org.apache.flink.streaming.connectors.pulsar.internal.PulsarDeserializer;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarFetcher;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarMetadataReader;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarRowFetcher;
 import org.apache.flink.streaming.runtime.tasks.ProcessingTimeService;
-import org.apache.flink.table.types.FieldsDataType;
-import org.apache.flink.table.types.utils.LegacyTypeInfoDataTypeConverter;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.ExceptionUtils;
 import org.apache.flink.util.SerializedValue;
@@ -32,10 +31,13 @@ import org.apache.flink.util.SerializedValue;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.pulsar.client.api.MessageId;
 import org.apache.pulsar.client.impl.conf.ClientConfigurationData;
+import org.apache.pulsar.common.schema.SchemaInfo;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
+
+import static org.apache.flink.streaming.connectors.pulsar.internal.PulsarOptions.USE_EXTEND_FIELD;
 
 /**
  * Emit Pulsar message as Row to Flink.
@@ -45,8 +47,16 @@ public class FlinkPulsarRowSource extends FlinkPulsarSource<Row> {
 
     private TypeInformation<Row> typeInformation;
 
+    public FlinkPulsarRowSource(String adminUrl, ClientConfigurationData clientConf, Properties properties, PulsarDeserializationSchema<Row> deserializer) {
+        super(adminUrl, clientConf, deserializer, properties);
+    }
+
     public FlinkPulsarRowSource(String adminUrl, ClientConfigurationData clientConf, Properties properties) {
         super(adminUrl, clientConf, (PulsarDeserializationSchema<Row>) null, properties);
+    }
+
+    public FlinkPulsarRowSource(String serviceUrl, String adminUrl, Properties properties,  PulsarDeserializationSchema<Row> deserializer) {
+        super(serviceUrl, adminUrl, deserializer, properties);
     }
 
     public FlinkPulsarRowSource(String serviceUrl, String adminUrl, Properties properties) {
@@ -55,18 +65,28 @@ public class FlinkPulsarRowSource extends FlinkPulsarSource<Row> {
 
     @Override
     public TypeInformation<Row> getProducedType() {
-        if (typeInformation == null) {
-            try (PulsarMetadataReader reader = new PulsarMetadataReader(adminUrl, clientConfigurationData, "", caseInsensitiveParams, -1, -1)) {
-                List<String> topics = reader.getTopics();
-                FieldsDataType schema = reader.getSchema(topics);
-                typeInformation = (TypeInformation<Row>) LegacyTypeInfoDataTypeConverter.toLegacyTypeInfo(schema);
-            } catch (Exception e) {
-                log.error("Failed to get schema for source with exception {}", ExceptionUtils.stringifyException(e));
-                typeInformation = null;
-            }
-        }
+        PulsarDeserializationSchema<Row> deserializer = getDeserializer();
+        return deserializer.getProducedType();
+    }
 
-        return typeInformation;
+    protected PulsarDeserializationSchema<Row> getDeserializer(){
+        if (deserializer != null){
+            return deserializer;
+        }
+        boolean useExtendField = Boolean.parseBoolean(properties.getProperty(USE_EXTEND_FIELD, "true"));
+        try (PulsarMetadataReader reader = new PulsarMetadataReader(adminUrl, clientConfigurationData, "", caseInsensitiveParams, -1, -1)) {
+            List<String> topics = reader.getTopics();
+            SchemaInfo pulsarSchema = reader.getPulsarSchema(topics);
+            synchronized (this) {
+                if (deserializer == null){
+                    deserializer = new PulsarDeserializer(pulsarSchema, null, useExtendField);
+                }
+            }
+            return deserializer;
+        } catch (Exception e) {
+            log.error("Failed to get schema for source with exception {}", ExceptionUtils.stringifyException(e));
+            throw new RuntimeException(e);
+        }
     }
 
     @Override
@@ -92,7 +112,7 @@ public class FlinkPulsarRowSource extends FlinkPulsarSource<Row> {
                 clientConfigurationData,
                 readerConf,
                 pollTimeoutMs,
-                null,
+                deserializer,
 				metadataReader);
     }
 }
