+++ pulsar-flink-1.9/src/main/java/org/apache/flink/streaming/connectors/pulsar/PulsarTableSource.java
@@ -18,8 +18,11 @@ import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.connectors.pulsar.config.StartupMode;
+import org.apache.flink.streaming.connectors.pulsar.internal.IncompatibleSchemaException;
 import org.apache.flink.streaming.connectors.pulsar.internal.PulsarMetadataReader;
+import org.apache.flink.streaming.connectors.pulsar.internal.SchemaTranslator;
 import org.apache.flink.streaming.connectors.pulsar.internal.SchemaUtils;
+import org.apache.flink.streaming.connectors.pulsar.internal.SimpleSchemaTranslator;
 import org.apache.flink.streaming.connectors.pulsar.internal.SourceSinkUtils;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.Types;
@@ -41,6 +44,7 @@ import org.apache.pulsar.client.admin.PulsarAdminException;
 import org.apache.pulsar.client.api.MessageId;
 import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.impl.conf.ClientConfigurationData;
+import org.apache.pulsar.common.schema.SchemaInfo;
 
 import java.util.Collections;
 import java.util.List;
@@ -79,6 +83,8 @@ public class PulsarTableSource
 
     private final TableSchema schema;
 
+    private final SchemaTranslator schemaTranslator;
+
     public PulsarTableSource(
             Optional<TableSchema> providedSchema,
             Optional<String> proctimeAttribute,
@@ -100,7 +106,7 @@ public class PulsarTableSource
 
         this.caseInsensitiveParams =
                 SourceSinkUtils.validateStreamSourceOptions(Maps.fromProperties(properties));
-
+        this.schemaTranslator = new SimpleSchemaTranslator(true);
         this.schema = inferTableSchema();
 
         this.proctimeAttribute = validateProctimeAttribute(proctimeAttribute);
@@ -170,9 +176,9 @@ public class PulsarTableSource
             try {
                 PulsarMetadataReader reader = new PulsarMetadataReader(adminUrl, new ClientConfigurationData(), "", caseInsensitiveParams, -1, -1);
                 List<String> topics = reader.getTopics();
-                FieldsDataType schema = reader.getSchema(topics);
-                return SchemaUtils.toTableSchema(schema);
-            } catch (PulsarClientException | PulsarAdminException | SchemaUtils.IncompatibleSchemaException e) {
+                SchemaInfo pulsarSchema = reader.getPulsarSchema(topics);
+                return schemaTranslator.pulsarSchemaToTableSchema(pulsarSchema);
+            } catch (PulsarClientException | PulsarAdminException | IncompatibleSchemaException e) {
                 log.error("Failed to fetch table schema", adminUrl);
                 throw new RuntimeException(e);
             }
