--- pulsar-flink-1.9/src/main/java/org/apache/flink/streaming/connectors/pulsar/internal/PulsarDeserializationSchema.java
@@ -1,49 +1,0 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.streaming.connectors.pulsar.internal;
-
-import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
-
-import org.apache.pulsar.client.api.Message;
-
-import java.io.IOException;
-import java.io.Serializable;
-
-/**
- * The deserialization schema describes how to turn the Pulsar reader messages
- * into data types (Java/Scala objects) that are processed by Flink.
- *
- * @param <T> The type created by the keyed deserialization schema.
- */
-public interface PulsarDeserializationSchema<T> extends Serializable, ResultTypeQueryable<T> {
-
-    /**
-     * Method to decide whether the element signals the end of the stream. If
-     * true is returned the element will not be emitted.
-     *
-     * @param nextElement The element to check for the end-of-stream signal.
-     * @return True, if the element signals end of stream, false otherwise.
-     */
-    boolean isEndOfStream(T nextElement);
-
-    /**
-     * Deserializes the Pulsar message.
-     *
-     * @param message Pulsar message to be deserialized.
-     * @return The deserialized message as an object.
-     * @throws IOException
-     */
-    T deserialize(Message message) throws IOException;
-}
