+++ pulsar-flink-1.9/src/test/java/org/apache/flink/streaming/connectors/pulsar/FlinkPulsarITest.java
@@ -40,6 +40,7 @@ import org.apache.flink.streaming.api.operators.StreamSink;
 import org.apache.flink.streaming.connectors.pulsar.internal.AvroDeser;
 import org.apache.flink.streaming.connectors.pulsar.internal.CachedPulsarClient;
 import org.apache.flink.streaming.connectors.pulsar.internal.JsonDeser;
+import org.apache.flink.streaming.connectors.pulsar.internal.PulsarDeserializationSchemaWrapper;
 import org.apache.flink.streaming.connectors.pulsar.internal.ReaderThread;
 import org.apache.flink.streaming.connectors.pulsar.testutils.FailingIdentityMapper;
 import org.apache.flink.streaming.connectors.pulsar.testutils.SingletonStreamSink;
@@ -149,7 +150,8 @@ public class FlinkPulsarITest extends PulsarTestBaseWithFlink {
             see.setParallelism(1);
 
             FlinkPulsarSource<String> source =
-                    new FlinkPulsarSource<String>("sev", "admin", new SimpleStringSchema(), props).setStartFromEarliest();
+                    new FlinkPulsarSource<String>("sev", "admin",
+                            new PulsarDeserializationSchemaWrapper<>(new SimpleStringSchema()), props).setStartFromEarliest();
 
             DataStream<String> stream = see.addSource(source);
             stream.print();
@@ -175,7 +177,8 @@ public class FlinkPulsarITest extends PulsarTestBaseWithFlink {
         props.setProperty("pulsar.reader.receiverQueueSize", "1000000");
 
         FlinkPulsarSource<Integer> source =
-                new FlinkPulsarSource<Integer>(serviceUrl, adminUrl, new IntegerDeserializer(), props)
+                new FlinkPulsarSource<Integer>(serviceUrl, adminUrl,
+                        new PulsarDeserializationSchemaWrapper<>(new IntegerDeserializer()), props)
                         .setStartFromEarliest();
 
         DataStream<Integer> stream = see.addSource(source);
@@ -379,7 +382,8 @@ public class FlinkPulsarITest extends PulsarTestBaseWithFlink {
         sourceProps.setProperty(TOPIC_SINGLE_OPTION_KEY, topic);
 
         FlinkPulsarSource<SchemaData.Foo> source =
-                new FlinkPulsarSource<>(serviceUrl, adminUrl, AvroDeser.of(SchemaData.Foo.class), sourceProps)
+                new FlinkPulsarSource<>(serviceUrl, adminUrl,
+                        new PulsarDeserializationSchemaWrapper<>(AvroDeser.of(SchemaData.Foo.class)), sourceProps)
                 .setStartFromEarliest();
 
         DataStream<Integer> ds = see.addSource(source)
@@ -415,7 +419,8 @@ public class FlinkPulsarITest extends PulsarTestBaseWithFlink {
         sourceProps.setProperty(TOPIC_SINGLE_OPTION_KEY, topic);
 
         FlinkPulsarSource<SchemaData.Foo> source =
-                new FlinkPulsarSource<>(serviceUrl, adminUrl, JsonDeser.of(SchemaData.Foo.class), sourceProps)
+                new FlinkPulsarSource<>(serviceUrl, adminUrl,
+                        new PulsarDeserializationSchemaWrapper<>(JsonDeser.of(SchemaData.Foo.class)), sourceProps)
                         .setStartFromEarliest();
 
         DataStream<Integer> ds = see.addSource(source)
@@ -736,7 +741,7 @@ public class FlinkPulsarITest extends PulsarTestBaseWithFlink {
             .subscriptionType(SubscriptionType.Exclusive)
             .subscribe();
 
-        env.addSource(new FlinkPulsarSource<>(serviceUrl, adminUrl, new SimpleStringSchema(), sourceProps)
+        env.addSource(new FlinkPulsarSource<>(serviceUrl, adminUrl, new PulsarDeserializationSchemaWrapper<>(new SimpleStringSchema()), sourceProps)
                 .setStartFromEarliest())
             .addSink(new FlinkPulsarSink<>(serviceUrl, adminUrl, Optional.of(outputTopic),
                 new Properties(), null, String.class))
